fix: clear pre-existing Windows clippy warnings (first --all-targets run on Win)
The workspace had only ever been linted on macOS (clippy 1.94). The first
Win11 run (clippy 1.95, --all-targets -D warnings) surfaced platform-cfg
warnings invisible on the Mac side. None relate to clipboard-files; fixing
here so the feature branch lints clean on both hosts before merge.

host (Windows-cfg code):
- injector.rs: drop redundant `as u16` on `scancode & 0xFF` (unnecessary_cast)
- ui/transfer_overlay.rs: `.build(&layout)` + non-mut binding (unnecessary_mut_passed)
- main.rs: fold `if is_port_combo` into the match arm guard (collapsible_match)
- session.rs: allow(dead_code) on the test-only has_shell (gated tests)

client (macOS-only code, dead on Win):
- app.rs, status_bar.rs, input/keymap.rs, keyboard_tap.rs: cfg_attr/allow
  dead_code on dock-icon fields, status-bar helpers, CGEventTap hotkey
  machinery
- main.rs: gate mem::forget(status_bar handle) to macOS (forget_non_drop —
  the non-macOS handle is an empty no-Drop placeholder)

Verified: cargo clippy native (1.94) + cross x86_64-pc-windows-gnu clean,
full workspace tests green (633 passed). The Rust 1.95 manual_checked_ops
case is handled in the preceding clipboard-files commit.

+++ apps/wiredesk-host/src/injector.rs
@@ -108,7 +108,7 @@ impl InputInjector for WindowsInjector {
 
         // Extended scancodes (0xE0xx) need KEYEVENTF_EXTENDEDKEY flag
         let (scan, flags) = if scancode & 0xFF00 == 0xE000 {
-            ((scancode & 0xFF) as u16, KEYEVENTF_SCANCODE | KEYEVENTF_EXTENDEDKEY)
+            (scancode & 0xFF, KEYEVENTF_SCANCODE | KEYEVENTF_EXTENDEDKEY)
         } else {
             (scancode, KEYEVENTF_SCANCODE)
         };
@@ -133,7 +133,7 @@ impl InputInjector for WindowsInjector {
         use windows::Win32::UI::Input::KeyboardAndMouse::*;
 
         let (scan, base_flags) = if scancode & 0xFF00 == 0xE000 {
-            ((scancode & 0xFF) as u16, KEYEVENTF_SCANCODE | KEYEVENTF_EXTENDEDKEY)
+            (scancode & 0xFF, KEYEVENTF_SCANCODE | KEYEVENTF_EXTENDEDKEY)
         } else {
             (scancode, KEYEVENTF_SCANCODE)
         };
